Rename physical query to clickhouse query
