+++ snuba/clickhouse/astquery.py
@@ -4,7 +4,7 @@ from snuba import settings
 from snuba.clickhouse.formatter import ClickhouseQueryFormatter
 from snuba.clickhouse.formatter import ClickhouseExpressionFormatter
 from snuba.query.parsing import ParsingContext
-from snuba.query.physical import Query
+from snuba.clickhouse.query import Query
 from snuba.request.request_settings import RequestSettings
