+++ snuba/clickhouse/dictquery.py
@@ -4,7 +4,7 @@ from snuba.clickhouse.formatter import ClickhouseQueryFormatter
 from snuba.datasets.dataset import Dataset
 from snuba.query.columns import column_expr, conditions_expr
 from snuba.query.parsing import ParsingContext
-from snuba.query.physical import Query
+from snuba.clickhouse.query import Query
 from snuba.request.request_settings import RequestSettings
