+++ snuba/clickhouse/query.py
@@ -1,5 +1,5 @@
 from snuba.query.logical import Query as LogicalQuery
 
-# This defines the type for the physical query that operates on storages.
+# This defines the type for the Clickhouse query that operates on storages.
 # Obviously this is supposed to become more complex than just a type alias.
 Query = LogicalQuery
