+++ snuba/query/processors/transaction_column_processor.py
@@ -1,5 +1,5 @@
 from snuba.query.expressions import Column, Expression, FunctionCall, Literal
-from snuba.query.physical import Query
+from snuba.clickhouse.query import Query
 from snuba.query.processors.physical import QueryProcessor
 from snuba.request.request_settings import RequestSettings
 
@@ -9,9 +9,7 @@ class TransactionColumnProcessor(QueryProcessor):
     Strip any dashes out of the event ID to match what is stored internally.
     """
 
-    def process_query(
-        self, query: Query, request_settings: RequestSettings
-    ) -> None:
+    def process_query(self, query: Query, request_settings: RequestSettings) -> None:
         def process_column(exp: Expression) -> Expression:
             if isinstance(exp, Column):
                 if exp.column_name == "event_id":
